chore: Move files to avoid 2 similar functions in same (split) package
