+++ OldKotlin/realm-models/src/main/kotlin/com/infomaniak/mail/data/models/correspondent/Recipient.kt
@@ -18,8 +18,8 @@
 package com.infomaniak.mail.data.models.correspondent
 
 import android.os.Parcel
-import com.infomaniak.core.common.extensions.customReadBoolean
-import com.infomaniak.core.common.extensions.customWriteBoolean
+import com.infomaniak.mail.internal.customReadBoolean
+import com.infomaniak.mail.internal.customWriteBoolean
 import com.infomaniak.mail.data.models.InternalModelProperties
 import io.realm.kotlin.types.EmbeddedRealmObject
 import io.realm.kotlin.types.annotations.Ignore
